Refactoring of import statements
