+++ src/citygml/core/object/building.py
@@ -1,5 +1,5 @@
-from citydpc.core.obejct.abstractBuilding import AbstractBuilding
-from citydpc.core.obejct.buildingPart import BuildingPart
+from citygml.core.object.abstractBuilding import AbstractBuilding
+from citygml.core.object.buildingPart import BuildingPart
 
 
 class Building(AbstractBuilding):
