+++ src/citygml/core/object/fileUtil.py
