+++ src/sim/write_energy_ade.py
